Enable C++11 and testing
Switching to C++11 involved switching from Boost.Tuple to the standard
tuple class, but that meant providing an operator<< that could print
standard tuples. Unfortunately, that involved adding a new operator<<
overload to namespace std, which isn't allowed. Until we can find a
solution to that, we'll just live with it. (The solution might involve
switching to a different testing framework that provides a documented
facility for printing otherwise unprintable types.)