Factor out, test, and fix get_float_type
When there are two different interpretations for a given float type,
this code uses the more descriptive form. For example, some forms are
invalid on the 80386, but had meaning as NaN on 8086; this code would
interpret it as NaN, on the assumption that a user has encountered the
given number somewhere in the wild and wants to know what it might mean.
It probably had meaning wherever it was found, so let's show that here.

This change also simplifies the interface for float_traits, removing
some unused traits.