Add display-float command
This has been sitting uncommitted in the sandbox for two years. I don't
know whether it works. I just want to clear out uncommitted work in the
sandbox in preparation for moving to a new dev host.

Looks like a standalone program for printing floats as hexadecimal
literals. Maybe useful for testing.